requirements.txt
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.7
django-appconf==1.0.2
django-compressor==2.1.1
django-extra-views==0.6.4
django-filter==1.0.4
django-haystack==2.5.1
django-oscar==1.4
django-qiniu-storage==2.3.1
django-tables2==1.7.1
django-treebeard==4.1.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
factory-boy==2.8.1
Faker==0.7.17
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.18
Markdown==2.6.8
mock==2.0.0
olefile==0.44
pbr==3.0.1
phonenumbers==8.5.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
purl==1.3.1
python-dateutil==2.6.0
pytz==2017.2
qiniu==7.1.7
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.20
urllib3==1.22
